refactor pageview code to be more re-usable
