+++ api/models/pageviews.js
@@ -1,6 +1,11 @@
 // @flow
+import { padStart } from 'lodash';
 const { db } = require('./db');
-import type { PageViewType } from '../../shared/types';
+import type {
+  PageViewType,
+  PageViewData,
+  PageViewResult,
+} from '../../shared/types';
 
 type AddPageViewParam = {
   id: string,
@@ -24,17 +29,6 @@ export const addPageView = ({
     .run();
 };
 
-type GetPageViewParam = {
-  id: string,
-  pageviewType: PageViewType,
-};
-export const getPageViews = ({ id, pageviewType }: GetPageViewParam) => {
-  return db
-    .table('pageviews')
-    .filter({ refId: id, refType: pageviewType.toLowerCase() })
-    .run();
-};
-
 /**
  * Gets the pageviews in an aggregated form
  */
@@ -44,17 +38,17 @@ type GetAggregatedViewsParams = {
   startDate?: string,
   endDate?: string,
 };
-type GetAggregatedViewsReturn = {
-  group: [number, number, number],
-  reduction: number,
-};
-export const getAggregatedViews = ({
+export const getAggregatedViews = async ({
   id,
   pageviewType,
   startDate,
   endDate,
-}: GetAggregatedViewsParams): Promise<GetAggregatedViewsReturn[]> => {
-  return db
+}: GetAggregatedViewsParams): Promise<PageViewResult> => {
+  type Result = {
+    group: [number, number, number],
+    reduction: number,
+  };
+  const results: Result[] = await db
     .table('pageviews')
     .getAll([id, pageviewType.toLowerCase()], { index: 'refId_refType' })
     .group([
@@ -64,4 +58,22 @@ export const getAggregatedViews = ({
     ])
     .count()
     .run();
+  const pageViewData = results.map(toPageViewData);
+  return {
+    resolution: 'day',
+    data: pageViewData,
+  };
 };
+
+/**
+ * Converts data returned by the data to PageViewData format
+ */
+function toPageViewData({ group, reduction }): PageViewData {
+  const [year, month, day] = group;
+  const paddedMonth = padStart(`${month}`, 2, '0');
+  const paddedDay = padStart(`${day}`, 2, '0');
+  return {
+    date: `${year}/${paddedMonth}/${paddedDay}`,
+    views: reduction,
+  };
+}
